Refactoring Code and adding Login View
