+++ Flutter/PlantArm/plantarmapp/lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:firebase_core/firebase_core.dart';
-import 'package:firebase_auth/firebase_auth.dart';
-import 'firebase_options.dart';
+import 'package:plantarmapp/view/login_view.dart';
+import 'package:plantarmapp/view/register_view.dart';
+
 
 void main() {
   WidgetsFlutterBinding.ensureInitialized;
@@ -11,97 +11,6 @@ void main() {
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),
-      home: const HomePage(),
+      home: const LoginView(),
     ),);
-}
-
-class HomePage extends StatefulWidget {
-  const HomePage({Key? key}) : super(key: key);
-
-  @override
-  State<HomePage> createState() => _HomePageState();
-}
-
-class _HomePageState extends State<HomePage> {
-
-  late final TextEditingController _email;
-  late final TextEditingController _password;
-
-  @override
-  void initState() {
-    // TODO: implement initState
-    _email = TextEditingController();
-    _password = TextEditingController();
-    super.initState();
-  }
-
-  @override
-  void dispose() {
-    // TODO: implement dispose
-    _email.dispose();
-    _password.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('PlantArm'),
-        backgroundColor: const Color.fromARGB(255, 0, 45, 0),
-        foregroundColor: const Color.fromARGB(255, 0, 255, 0),
-      ),
-      body: FutureBuilder(
-        future: Firebase.initializeApp(
-          options: DefaultFirebaseOptions.currentPlatform
-        ),
-        builder: (context, snapshot){
-
-          switch (snapshot.connectionState){
-            
-            case ConnectionState.done:
-              return Column(
-                children: [
-                  TextField(
-                    controller: _email,
-                    enableSuggestions: false,
-                    autocorrect: false,
-                    keyboardType: TextInputType.emailAddress,
-                    decoration: const InputDecoration(
-                      hintText: 'Email',
-                      hintStyle: TextStyle(color: Colors.black),
-                    ),
-                  ),
-                  TextField(
-                    controller: _password,
-                    enableSuggestions: false,
-                    autocorrect: false,
-                    obscureText: true,
-                    decoration: const InputDecoration(
-                      hintText: 'Password',
-                      hintStyle: TextStyle(color: Colors.black),
-                    ),
-                  ),
-                  TextButton(onPressed: (() async {
-                    final username = _email.text;
-                    final password = _password.text;
-                    final userCredential = await FirebaseAuth.instance.createUserWithEmailAndPassword(
-                      email: username,
-                      password: password);
-                      print(userCredential);
-                  }), style: ButtonStyle(
-                        backgroundColor: MaterialStateProperty.all(const Color.fromARGB(255, 0, 45, 0)),
-                        foregroundColor: MaterialStateProperty.all(const Color.fromARGB(255, 0, 255, 0))),
-                      child: 
-                        const Text('Register')),
-                ],
-              );
-              default:
-                return const Text('Loading...');
-
-          }
-        }
-      )
-    );
-  }
 }
